Flatten the class hierarchy for SA token fetching
The TokenClient was not actually client and did not provide
any value on its own. This commit takes out the only function
of it that had any value and extracts it outside, taking the
tokenManager as an argument rather than wrapping it in another
object.
