+++ pkg/token/token_manager.go
@@ -25,6 +25,7 @@ package token
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"math/rand"
@@ -197,4 +198,50 @@ func keyFunc(name, namespace string, tr *authenticationv1.TokenRequest) string {
 	}
 
 	return fmt.Sprintf("%q/%q/%#v/%#v/%#v", name, namespace, tr.Spec.Audiences, exp, ref)
+}
+
+// SecretProviderServiceAccountTokenAttrs returns the token for the federated service account that can be bound to the pod.
+// This token will be sent to the providers and is of the format:
+//
+//	"csi.storage.k8s.io/serviceAccount.tokens": {
+//	  <audience>: {
+//	    'token': <token>,
+//	    'expirationTimestamp': <expiration timestamp in RFC3339 format>,
+//	  },
+//	  ...
+//	}
+//
+// ref: https://kubernetes-csi.github.io/docs/token-requests.html#usage
+func SecretProviderServiceAccountTokenAttrs(tokenManager *Manager, namespace, serviceAccountName string, audiences []string) (map[string]string, error) {
+	if len(audiences) == 0 {
+		return nil, nil
+	}
+
+	outputs := map[string]authenticationv1.TokenRequestStatus{}
+	var tokenExpirationSeconds int64 = 600
+
+	for _, aud := range audiences {
+		tr := &authenticationv1.TokenRequest{
+			Spec: authenticationv1.TokenRequestSpec{
+				ExpirationSeconds: &tokenExpirationSeconds,
+				Audiences:         []string{aud},
+			},
+		}
+
+		tr, err := tokenManager.GetServiceAccountToken(namespace, serviceAccountName, tr)
+		if err != nil {
+			return nil, err
+		}
+		outputs[aud] = tr.Status
+	}
+
+	klog.V(5).InfoS("Fetched service account token attrs", "serviceAccountName", serviceAccountName, "namespace", namespace)
+	tokens, err := json.Marshal(outputs)
+	if err != nil {
+		return nil, err
+	}
+
+	return map[string]string{
+		"csi.storage.k8s.io/serviceAccount.tokens": string(tokens),
+	}, nil
 }
